# src/test/resources/contracts/api-contract.yaml
openapi: 3.0.3
info:
  title: AdBoard API Contract
  version: 1.0.0
  description: Full API contract for manual, integration, and contract testing.
servers:
  - url: http://localhost:8081
...
paths:
  /v1/auth/register:
    post:
      tags: [Auth]
      summary: Register user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequestRegistration'
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseRegistration'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
  /v1/auth/login:
    post:
      tags: [Auth]
      summary: Login user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid credentials or request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
  /v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh tokens
      operationId: refreshTokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Refresh successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
  /v1/auth/logout:
    post:
      tags: [Auth]
      summary: Logout user
      operationId: logoutUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '204':
          description: Logout successful
        '400':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    UserRequestRegistration:
      type: object
      required: [name, email, phone, password]
      properties:
        name:
          type: string
          minLength: 1
          example: Ivan Ivanov
        email:
          type: string
          format: email
          example: ivan@example.com
        phone:
          type: string
          minLength: 1
          example: "+79990000000"
        password:
          type: string
          minLength: 1
          example: secret
    UserResponseRegistration:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Ivan Ivanov
        email:
          type: string
          format: email
          example: ivan@example.com
        phone:
          type: string
          example: "+79990000000"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: ivan@example.com
        password:
          type: string
          minLength: 1
          example: secret
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 1
          example: eyJhbGciOi...
    LogoutRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 1
          example: eyJhbGciOi...
    LoginResponse:
      type: object
      required:
        - access_token
        - refresh_token
        - expires_in
        - refresh_expires_in
        - token_type
      properties:
        access_token:
          type: string
          example: eyJhbGciOi...
        refresh_token:
          type: string
          example: eyJhbGciOi...
        expires_in:
          type: integer
          format: int64
          example: 300
        refresh_expires_in:
          type: integer
          format: int64
          example: 1800
        token_type:
          type: string
          example: Bearer
...
    ExceptionResponse:
      type: object
      properties:
        message:
          type: string
          example: Advertisement with id 1 not found
        status:
          type: integer
          example: 404
        timestamp:
          type: string
          format: date-time
          example: '2026-04-14T19:00:00'